Generate random samples from a primary event censored distribution

This function generates random samples from a primary event censored distribution. It adjusts the distribution by accounting for the primary event distribution and potential truncation at a maximum delay (D). The function allows for custom primary event distributions and delay distributions.

Usage

rprimarycensoreddist(
  n,
  rdist,
  pwindow = 1,
  swindow = 1,
  D = Inf,
  rprimary = stats::runif,
  rprimary_args = list(),
  oversampling_factor = 1.2,
  ...
)

rpcens(
  n,
  rdist,
  pwindow = 1,
  swindow = 1,
  D = Inf,
  rprimary = stats::runif,
  rprimary_args = list(),
  oversampling_factor = 1.2,
  ...
)

Arguments

n

Number of random samples to generate.

rdist

Function to generate random samples from the delay distribution for example stats::rlnorm() for lognormal distribution.

pwindow

Primary event window

swindow

Secondary event window (default: 1)

D

Maximum delay (truncation point). If finite, the distribution is truncated at D. If set to Inf, no truncation is applied. Defaults to Inf.

rprimary

Function to generate random samples from the primary distribution (default is stats::runif()).

rprimary_args

List of additional arguments to be passed to rprimary.

oversampling_factor

Factor by which to oversample the number of samples to account for truncation (default is 1.2).

...

Additional arguments to be passed to the distribution function.

Value

Vector of random samples from the primary event censored distribution

Details

The primary event censored random samples are generated by first oversampling to account for potential truncation. Primary event times are generated from the specified primary event distribution, and delays are generated from the specified delay distribution. The total delays are calculated by adding the primary event times and the delays. These total delays are then rounded to the nearest secondary event window (swindow). Truncation is applied to ensure that the delays are within the specified range [0, D]. If the number of valid samples is less than the desired number of samples (n), additional samples are generated until the required number of valid samples is obtained.

See also

Examples

# Example: Lognormal distribution with uniform primary events
rprimarycensoreddist(10, rlnorm, meanlog = 0, sdlog = 1)
#>  [1] 1 3 3 1 1 6 1 3 4 1

# Example: Lognormal distribution with exponential growth primary events
rprimarycensoreddist(
  10, rlnorm,
  rprimary = rexpgrowth, rprimary_args = list(r = 0.2),
  meanlog = 0, sdlog = 1
)
#>  [1] 1 3 1 0 1 1 2 9 1 1